backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-multipart
python-dotenv
openai
prokerala-api